% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{resource_tracker}
\alias{resource_tracker}
\title{Raw access to the resource_tracker Python module.}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 1.
}
\usage{
resource_tracker
}
\description{
If you don't need advanced features, use \code{\link{ResourceTracker}} instead.
}
\keyword{datasets}
