% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResorceTracker.R
\name{ResourceTracker}
\alias{ResourceTracker}
\title{Resource Tracker Object}
\description{
And R-friendly wrapper around the Python \code{resource-tracker} package's
\code{ResourceTracker} class.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{pid}}{The process ID of the tracked process.}

\item{\code{system_metrics}}{The system metrics of the tracked process.}

\item{\code{process_metrics}}{The process metrics of the tracked process.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResourceTracker-new}{\code{ResourceTracker$new()}}
\item \href{#method-ResourceTracker-clone}{\code{ResourceTracker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceTracker-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceTracker-new}{}}}
\subsection{Method \code{new()}}{
Initialize the ResourceTracker object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceTracker$new(pid = Sys.getpid())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pid}}{The process ID to track. Defaults to current process ID.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceTracker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceTracker-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceTracker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
