% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceTracker.R
\name{ResourceTracker}
\alias{ResourceTracker}
\title{Resource Tracker Object}
\description{
And R-friendly wrapper around the Python \code{resource-tracker} package's
\code{ResourceTracker} class.
}
\examples{
tracker <- ResourceTracker$new()
tracker$wait_for_samples(1)
tracker$stats()
tracker$recommend_resources()
tracker$recommend_server()

\dontrun{
tracker$report()
}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{running}}{(logical) Whether the resource tracker is running.}

\item{\code{pid}}{(integer) The process ID of the tracked process.}

\item{\code{n_samples}}{(integer) The number of samples taken.}

\item{\code{system_metrics}}{(data.frame) The system metrics of the tracked process.}

\item{\code{process_metrics}}{(data.frame) The process metrics of the tracked process.}

\item{\code{cloud_info}}{(list) The cloud environment discovered, including the cloud provider, instance type, and the datacenter region.}

\item{\code{server_info}}{(list) The server specs discovered, including the
operating system, number of vCPUs, memory amount, GPUs and VRAM, and a
guess if the server is dedicated to the tracked process(es) or shared
with other processes.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResourceTracker-new}{\code{ResourceTracker$new()}}
\item \href{#method-ResourceTracker-start}{\code{ResourceTracker$start()}}
\item \href{#method-ResourceTracker-stop}{\code{ResourceTracker$stop()}}
\item \href{#method-ResourceTracker-wait_for_samples}{\code{ResourceTracker$wait_for_samples()}}
\item \href{#method-ResourceTracker-get_combined_metrics}{\code{ResourceTracker$get_combined_metrics()}}
\item \href{#method-ResourceTracker-stats}{\code{ResourceTracker$stats()}}
\item \href{#method-ResourceTracker-recommend_resources}{\code{ResourceTracker$recommend_resources()}}
\item \href{#method-ResourceTracker-recommend_server}{\code{ResourceTracker$recommend_server()}}
\item \href{#method-ResourceTracker-report}{\code{ResourceTracker$report()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceTracker-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceTracker-new}{}}}
\subsection{Method \code{new()}}{
Initialize the ResourceTracker object in the background.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceTracker$new(
  pid = Sys.getpid(),
  children = TRUE,
  interval = 1,
  method = NULL,
  autostart = TRUE,
  track_processes = TRUE,
  track_system = TRUE,
  discover_server = TRUE,
  discover_cloud = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pid}}{The process ID to track. Defaults to current process ID.}

\item{\code{children}}{Whether to track child processes. Defaults to True.}

\item{\code{interval}}{Sampling interval in seconds. Defaults to 1.}

\item{\code{method}}{Multiprocessing method. Defaults to None, which tries to fork on Linux and macOS, and spawn on Windows.}

\item{\code{autostart}}{Whether to start tracking immediately. Defaults to True.}

\item{\code{track_processes}}{Whether to track resource usage at the process level. Defaults to True.}

\item{\code{track_system}}{Whether to track system-wide resource usage. Defaults to True.}

\item{\code{discover_server}}{Whether to discover the server specs in the background at startup. Defaults to True.}

\item{\code{discover_cloud}}{Whether to discover the cloud environment in the background at startup. Defaults to True.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The ResourceTracker object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceTracker-start"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceTracker-start}{}}}
\subsection{Method \code{start()}}{
Start tracking resource usage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceTracker$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceTracker-stop"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceTracker-stop}{}}}
\subsection{Method \code{stop()}}{
Stop tracking resource usage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceTracker$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceTracker-wait_for_samples"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceTracker-wait_for_samples}{}}}
\subsection{Method \code{wait_for_samples()}}{
Wait for a certain number of samples to be collected.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceTracker$wait_for_samples(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{The number of samples to wait for. Defaults to 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceTracker-get_combined_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceTracker-get_combined_metrics}{}}}
\subsection{Method \code{get_combined_metrics()}}{
Get the combined metrics of the tracked process.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceTracker$get_combined_metrics(
  bytes = FALSE,
  human_names = FALSE,
  system_prefix = NULL,
  process_prefix = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bytes}}{Whether to return the metrics in bytes. Defaults to False.}

\item{\code{human_names}}{Whether to return the metrics in human readable names. Defaults to False.}

\item{\code{system_prefix}}{The prefix to add to the system metrics. Defaults to NULL.}

\item{\code{process_prefix}}{The prefix to add to the process metrics. Defaults to NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of the combined metrics.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceTracker-stats"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceTracker-stats}{}}}
\subsection{Method \code{stats()}}{
Get the stats of the tracked process.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceTracker$stats()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{specs}}{The specs to get the stats for. Defaults to NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame of the stats.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceTracker-recommend_resources"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceTracker-recommend_resources}{}}}
\subsection{Method \code{recommend_resources()}}{
Recommend hardware resources for the tracked process.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceTracker$recommend_resources()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of the recommended resources.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceTracker-recommend_server"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceTracker-recommend_server}{}}}
\subsection{Method \code{recommend_server()}}{
Recommend a cloud server for the tracked process.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceTracker$recommend_server()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list including the recommended server specs, vendor, and the estimated cost.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceTracker-report"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceTracker-report}{}}}
\subsection{Method \code{report()}}{
Generate an interactive HTML report of the resource usage and recommendations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceTracker$report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A Report object with the \code{save} (to file) and \code{browse} (to open in browser) methods. Convert to string with \code{as.character} if needed.
}
}
}
